+++ ICD.Connect.Settings/SPlusShims/ISplusShim.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace ICD.Connect.Settings.SPlusShims
+{
+	public interface ISPlusShim : IDisposable
+	{
+		 string Location { get; set; }
+	}
+}