+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Added interface for username/password configuration
+ - Added SPlusGlobalEvents framework, for initializing from s+ and future s+ eventing.
 
 ### Changed
  - Moving originators and settings classes into subdirectories